feat: add metadata and optimization to projects page
