+++ src/components/ui/Download.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import Button from './Button';
 
 interface DownloadProps {
